+++ CHANGES.rst
@@ -10,6 +10,7 @@ Fixed
 ^^^^^
 - Locked account cannot reset their password. :issue:`250`
 - Canaille has default admin ACLs. :issue:`253`
+- SMS edition form. :issue:`259`
 
 [0.0.68] - 2025-03-18
 ---------------------