+++ CHANGES.rst
@@ -1,3 +1,11 @@
+[0.0.70] - Unreleased
+---------------------
+
+Added
+^^^^^
+- Add a :attr:`~canaille.core.configuration.CoreSettings.LOGIN_ATTRIBUTES` configuration
+  parameter to allow to disable https forced redirections.
+
 [0.0.69] - 2025-03-28
 ---------------------